Alternative fix for trusty (to work around bug).
So I found that upon libvirt-bin restart there was a bug causing the
bridge information to go away. This led me to an alternative method
that instead pushes a flag (2) to get inactive domain XML instead of
current running XML. Since we never changes these this is best for
our use cases.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>